# docs/api.swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome to the ORY Hydra HTTP API documentation. You will find documentation for all HTTP APIs here.",
    "title": "ORY Hydra",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
...
    "/oauth2/auth": {
      "get": {
        "description": "This endpoint is not documented here because you should never use your own implementation to perform OAuth2 flows.\nOAuth2 is a very popular protocol and a library for your programming language will exists.\n\nTo learn more about this flow please refer to the specification: https://tools.ietf.org/html/rfc6749",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "public"
        ],
        "summary": "The OAuth 2.0 authorize endpoint",
        "operationId": "oauthAuth",
        "responses": {
          "302": {
            "$ref": "#/responses/emptyResponse"
          },
          "401": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
    "/oauth2/token": {
      "post": {
        "security": [
          {
            "basic": []
          },
          {
            "oauth2": []
          }
        ],
        "description": "The client makes a request to the token endpoint by sending the\nfollowing parameters using the \"application/x-www-form-urlencoded\" HTTP\nrequest entity-body.\n\n\u003e Do not implement a client for this endpoint yourself. Use a library. There are many libraries\n\u003e available for any programming language. You can find a list of libraries here: https://oauth.net/code/\n\u003e\n\u003e Do not the the Hydra SDK does not implement this endpoint properly. Use one of the libraries listed above!",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "public"
        ],
        "summary": "The OAuth 2.0 token endpoint",
        "operationId": "oauth2Token",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "GrantType",
            "name": "grant_type",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Code",
            "name": "code",
            "in": "formData"
          },
          {
            "type": "string",
            "x-go-name": "RedirectURI",
            "name": "redirect_uri",
            "in": "formData"
          },
          {
            "type": "string",
            "x-go-name": "ClientID",
            "name": "client_id",
            "in": "formData"
          }
        ],
        "responses": {
          "200": {
            "description": "oauth2TokenResponse",
            "schema": {
              "$ref": "#/definitions/oauth2TokenResponse"
            }
          },
          "401": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "genericError": {
      "description": "Error responses are sent when an error (e.g. unauthorized, bad request, ...) occurred.",
      "type": "object",
      "title": "Error response",
      "required": [
        "error"
      ],
      "properties": {
        "debug": {
          "description": "Debug contains debug information. This is usually not available and has to be enabled.",
          "type": "string",
          "x-go-name": "Debug",
          "example": "The database adapter was unable to find the element"
        },
        "error": {
          "description": "Name is the error name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "The requested resource could not be found"
        },
        "error_description": {
          "description": "Description contains further information on the nature of the error.",
          "type": "string",
          "x-go-name": "Description",
          "example": "Object with ID 12345 does not exist"
        },
        "status_code": {
          "description": "Code represents the error status code (404, 403, 401, ...).",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 404
        }
      },
      "x-go-package": "github.com/tyaps/hydra"
    },
...
    "oauth2TokenResponse": {
      "description": "The Access Token Response",
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string",
          "x-go-name": "AccessToken"
        },
        "expires_in": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExpiresIn"
        },
        "refresh_token": {
          "type": "string",
          "x-go-name": "RefreshToken"
        },
        "token_type": {
          "type": "string",
          "x-go-name": "TokenType"
        }
      },
      "x-go-name": "swaggeroauth2TokenResponse",
      "x-go-package": "github.com/tyaps/hydra/oauth2"
    },
...
  },
  "responses": {
    "emptyResponse": {
      "description": "Empty responses are sent when, for example, resources are deleted. The HTTP status code for empty responses is\ntypically 201."
    },
...
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    },
    "oauth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "/oauth2/auth",
      "tokenUrl": "/oauth2/token",
      "scopes": {
        "offline": "A scope required when requesting refresh tokens (alias for `offline`)",
        "offline_access": "A scope required when requesting refresh tokens",
        "openid": "Request an OpenID Connect ID Token"
      }
    }
  },
...
}